# openapi.yaml
openapi: 3.0.3
info:
  title: traPhoto(仮) API
  description: 画像とアルバムを管理するためのAPI
  version: 1.0.0
  contact:
    name: traP-jp
    url: https://github.com/traP-jp/1m25_10
servers:
  - url: http://localhost:3001
    description: 開発サーバー（モック）
  - url: https://api.your-domain.com
    description: 本番サーバー
paths:
...
  /api/auth/request:
    get:
      summary: traQ OAuth リクエスト
      description: ブラウザをtraQの認可画面へリダイレクトします（サーバー側でstate/code_verifierを一時Cookieに保存）
      operationId: authRequest
      tags:
        - Auth
      parameters:
        - name: callback
          in: query
          required: false
          description: '認証後に戻る相対パス（例: "/account"）'
          schema:
            type: string
      responses:
        '302':
          description: 認可サーバーへリダイレクト
          headers:
            Location:
              description: traQ 認可エンドポイント
              schema:
                type: string
        '500':
          $ref: '#/components/responses/ServerError'
  /api/auth/callback:
    get:
      summary: traQ OAuth コールバック
      description: 認可コードをアクセストークンに交換し、クッキーへ保存後にフロントへリダイレクト
      operationId: authCallback
      tags:
        - Auth
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: フロントエンドへリダイレクト（Set-Cookie でセッション用トークンを設定）
          headers:
            Location:
              description: 'リダイレクト先（callback 指定があればそれ、なければ"/"）'
              schema:
                type: string
        '400':
          description: code/state が不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: 認可サーバーとのトークン交換に失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/logout:
    post:
      summary: ログアウト
      description: セッションCookieを破棄します。
      operationId: authLogout
      tags:
        - Auth
      responses:
        '204':
          description: 成功（内容なし）
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: エラーメッセージ
          example: "Resource not found"
        message:
          type: string
          description: 追加のエラー詳細
          example: "The requested resource could not be found"
  responses:
...
    ServerError:
      description: 内部サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: traq-auth-token
      description: traQ OAuth 交換後にサーバーがセットするアクセストークン Cookie
...